backend/requirements.txt
Django>=2.2.0,<2.3.0
djangorestframework>=3.11.0,<3.12.0
django-model-utils>=4.0.0,<4.1.0
psycopg2-binary>=2.8.0,<2.9.0
redis>=3.4.0,<3.5.0
djoser>=2.0.0,<2.1.0